added: PopUpEditText.java, user_edittext_popup.xml modified: UserTab, user_tab.xml, strings.xml
PopupEditText is a class for holding popup window and logic. constructor instantiates popup view with edittext and button inside that layout

userTab calls PopupEditText constructor when ButtonAddUser is clicked.

future: need logic for both buttonadduser and buttonedittext, need to get edittext and store for use in db query, need to display the user in the list view upon creation